+++ src/components/ConnectWallet/ConnectWalletDialog.tsx
@@ -8,7 +8,7 @@ import Column from 'components/Column'
 import { Header } from 'components/Dialog'
 import Row from 'components/Row'
 import EventEmitter from 'events'
-import { connections, useConnectCallback, Web3Connection } from 'hooks/connectWeb3/useActiveWeb3React'
+import { connections, useConnectCallback, Web3Connection } from 'hooks/connectWeb3/useWeb3React'
 import { atom, useAtom } from 'jotai'
 import QRCode from 'qrcode'
 import { useEffect, useState } from 'react'
