+++ src/components/ConnectWallet/index.tsx
@@ -1,5 +1,5 @@
 import { useWeb3React } from '@web3-react/core'
-import { connections } from 'hooks/connectWeb3/useActiveWeb3React'
+import { connections } from 'hooks/connectWeb3/useWeb3React'
 import { useEffect } from 'react'
 
 import ConnectedWalletChip from './ConnectedWalletChip'
@@ -13,7 +13,7 @@ interface WalletProps {
 export default function Wallet({ disabled, onClickConnectWallet }: WalletProps) {
   // Attempt to connect eagerly on mount
   useEffect(() => {
-    connections.forEach(([wallet, _]) => wallet?.connectEagerly?.())
+    connections.forEach(([wallet, _]) => (wallet?.connectEagerly ? wallet?.connectEagerly() : wallet?.activate()))
   }, [])
 
   const { account, isActive } = useWeb3React()
