+++ src/components/Widget.tsx
@@ -2,7 +2,7 @@ import { JsonRpcProvider } from '@ethersproject/providers'
 import { TokenInfo } from '@uniswap/token-lists'
 import { Provider as Eip1193Provider } from '@web3-react/types'
 import { DEFAULT_LOCALE, SUPPORTED_LOCALES, SupportedLocale } from 'constants/locales'
-import { ActiveWeb3Provider } from 'hooks/connectWeb3/useActiveWeb3React'
+import { ActiveWeb3Provider } from 'hooks/connectWeb3/useWeb3React'
 import { TransactionsUpdater } from 'hooks/transactions'
 import { BlockNumberProvider } from 'hooks/useBlockNumber'
 import { TokenListProvider } from 'hooks/useTokenList'
